pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tracemap"
version = "0.1.0"
description = "Visual data attribution for image classification."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Amit" },
]
dependencies = [
    "numpy>=1.26",
    "pillow>=10.2",
    "torch>=2.2",
    "torchvision>=0.17",
]

[project.optional-dependencies]
demo = [
    "streamlit>=1.33",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"
